Changed IStratumServerConfig.cs:Diff, VardiffConfig.cs:MinDiff and MaxDiff to a float so X11 coins can be used now too. Fixed a bug within StratumMiner.cs where miner default difficulty was not using configuration based value (fixes #313).

+++ src/CoiniumServ/Server/Mining/Stratum/Config/IStratumServerConfig.cs
@@ -21,7 +21,6 @@
 // 
 #endregion
 
-using System;
 using CoiniumServ.Vardiff;
 
 namespace CoiniumServ.Server.Mining.Stratum.Config
@@ -31,7 +30,7 @@ namespace CoiniumServ.Server.Mining.Stratum.Config
         /// <summary>
         /// default difficulty assigned to newly connected miners.
         /// </summary>
-        Int32 Diff { get; }
+        float Diff { get; }
 
         /// <summary>
         /// vardiff configuration.
